List available bundles for a dataset

Returns the full names of all white-matter bundles available in the specified dataset. The returned names are the values to supply to the bundle argument of import_bundle().

Usage

available_bundles(dataset)

Arguments

dataset

A string naming the dataset. Must be one of the values returned by available_datasets().

Value

A character vector of bundle names.

Examples

available_bundles("HCP1065")
#>  [1] "Anterior Commissure"                       
#>  [2] "Left Arcuate Fasciculus"                   
#>  [3] "Right Arcuate Fasciculus"                  
#>  [4] "Left Acoustic Radiation"                   
#>  [5] "Right Acoustic Radiation"                  
#>  [6] "Left Frontal Parietal Cingulum"            
#>  [7] "Right Frontal Parietal Cingulum"           
#>  [8] "Left Frontal Parahippocampal Cingulum"     
#>  [9] "Right Frontal Parahippocampal Cingulum"    
#> [10] "Left Parahippocampal Cingulum"             
#> [11] "Right Parahippocampal Cingulum"            
#> [12] "Left Parahippocampal Parietal Cingulum"    
#> [13] "Right Parahippocampal Parietal Cingulum"   
#> [14] "Left Parolfactory Cingulum"                
#> [15] "Right Parolfactory Cingulum"               
#> [16] "Left Cerebellum"                           
#> [17] "Right Cerebellum"                          
#> [18] "Left Corticobulbar Tract"                  
#> [19] "Right Corticobulbar Tract"                 
#> [20] "Corpus Callosum"                           
#> [21] "Left Cranial Nerve II"                     
#> [22] "Right Cranial Nerve II"                    
#> [23] "Left Cranial Nerve III"                    
#> [24] "Right Cranial Nerve III"                   
#> [25] "Left Cranial Nerve V"                      
#> [26] "Right Cranial Nerve V"                     
#> [27] "Left Cranial Nerve VII"                    
#> [28] "Right Cranial Nerve VII"                   
#> [29] "Left Cranial Nerve VIII"                   
#> [30] "Right Cranial Nerve VIII"                  
#> [31] "Left Frontal Corticopontine Tract"         
#> [32] "Right Frontal Corticopontine Tract"        
#> [33] "Left Occipital Corticopontine Tract"       
#> [34] "Right Occipital Corticopontine Tract"      
#> [35] "Left Parietal Corticopontine Tract"        
#> [36] "Right Parietal Corticopontine Tract"       
#> [37] "Left Anterior Corticostriatal Tract"       
#> [38] "Right Anterior Corticostriatal Tract"      
#> [39] "Left Posterior Corticostriatal Tract"      
#> [40] "Right Posterior Corticostriatal Tract"     
#> [41] "Left Superior Corticostriatal Tract"       
#> [42] "Right Superior Corticostriatal Tract"      
#> [43] "Left Corticospinal Tract"                  
#> [44] "Right Corticospinal Tract"                 
#> [45] "Left Dentatorubrothalamic Tract"           
#> [46] "Right Dentatorubrothalamic Tract"          
#> [47] "Left Extreme Capsule"                      
#> [48] "Right Extreme Capsule"                     
#> [49] "Left Fornix"                               
#> [50] "Right Fornix"                              
#> [51] "Left Frontal Aslant Tract"                 
#> [52] "Right Frontal Aslant Tract"                
#> [53] "Left Inferior Cerebellar Peduncle"         
#> [54] "Right Inferior Cerebellar Peduncle"        
#> [55] "Left Inferior Fronto-Occipital Fasciculus" 
#> [56] "Right Inferior Fronto-Occipital Fasciculus"
#> [57] "Left Inferior Longitudinal Fasciculus"     
#> [58] "Right Inferior Longitudinal Fasciculus"    
#> [59] "Middle Cerebellar Peduncle"                
#> [60] "Left Middle Longitudinal Fasciculus"       
#> [61] "Right Middle Longitudinal Fasciculus"      
#> [62] "Left Medial Lemniscus"                     
#> [63] "Right Medial Lemniscus"                    
#> [64] "Left Optic Radiation"                      
#> [65] "Right Optic Radiation"                     
#> [66] "Left Parietal Aslant Tract"                
#> [67] "Right Parietal Aslant Tract"               
#> [68] "Left Reticulospinal Tract"                 
#> [69] "Right Reticulospinal Tract"                
#> [70] "Superior Cerebellar Peduncle"              
#> [71] "Left Superior Longitudinal Fasciculus I"   
#> [72] "Right Superior Longitudinal Fasciculus I"  
#> [73] "Left Superior Longitudinal Fasciculus II"  
#> [74] "Right Superior Longitudinal Fasciculus II" 
#> [75] "Left Superior Longitudinal Fasciculus III" 
#> [76] "Right Superior Longitudinal Fasciculus III"
#> [77] "Left Anterior Thalamic Radiation"          
#> [78] "Right Anterior Thalamic Radiation"         
#> [79] "Left Posterior Thalamic Radiation"         
#> [80] "Right Posterior Thalamic Radiation"        
#> [81] "Left Superior Thalamic Radiation"          
#> [82] "Right Superior Thalamic Radiation"         
#> [83] "Left Uncinate Fasciculus"                  
#> [84] "Right Uncinate Fasciculus"                 
#> [85] "Vermis"                                    
#> [86] "Left Vertical Occipital Fasciculus"        
#> [87] "Right Vertical Occipital Fasciculus"       
available_bundles("TractSeg")
#>  [1] "Left Arcuate Fasciculus"                   
#>  [2] "Right Arcuate Fasciculus"                  
#>  [3] "Left Anterior Thalamic Radiation"          
#>  [4] "Right Anterior Thalamic Radiation"         
#>  [5] "Commissure Anterior"                       
#>  [6] "Corpus Callosum"                           
#>  [7] "Corpus Callosum Rostrum"                   
#>  [8] "Corpus Callosum Genu"                      
#>  [9] "Corpus Callosum Rostral Body"              
#> [10] "Corpus Callosum Anterior Midbody"          
#> [11] "Corpus Callosum Posterior Midbody"         
#> [12] "Corpus Callosum Isthmus"                   
#> [13] "Corpus Callosum Splenium"                  
#> [14] "Left Cingulum"                             
#> [15] "Right Cingulum"                            
#> [16] "Left Corticospinal Tract"                  
#> [17] "Right Corticospinal Tract"                 
#> [18] "Left Fronto-Pontine Tract"                 
#> [19] "Right Fronto-Pontine Tract"                
#> [20] "Left Fornix"                               
#> [21] "Right Fornix"                              
#> [22] "Left Inferior Cerebellar Peduncle"         
#> [23] "Right Inferior Cerebellar Peduncle"        
#> [24] "Left Inferior Fronto-Occipital Fasciculus" 
#> [25] "Right Inferior Fronto-Occipital Fasciculus"
#> [26] "Left Inferior Longitudinal Fasciculus"     
#> [27] "Right Inferior Longitudinal Fasciculus"    
#> [28] "Middle Cerebellar Peduncle"                
#> [29] "Left Middle Longitudinal Fasciculus"       
#> [30] "Right Middle Longitudinal Fasciculus"      
#> [31] "Left Optic Radiation"                      
#> [32] "Right Optic Radiation"                     
#> [33] "Left Parieto-Occipital Pontine Tract"      
#> [34] "Right Parieto-Occipital Pontine Tract"     
#> [35] "Left Superior Cerebellar Peduncle"         
#> [36] "Right Superior Cerebellar Peduncle"        
#> [37] "Left Superior Longitudinal Fasciculus I"   
#> [38] "Right Superior Longitudinal Fasciculus I"  
#> [39] "Left Superior Longitudinal Fasciculus II"  
#> [40] "Right Superior Longitudinal Fasciculus II" 
#> [41] "Left Superior Longitudinal Fasciculus III" 
#> [42] "Right Superior Longitudinal Fasciculus III"
#> [43] "Left Superior Thalamic Radiation"          
#> [44] "Right Superior Thalamic Radiation"         
#> [45] "Left Striato-Fronto-Orbital Tract"         
#> [46] "Right Striato-Fronto-Orbital Tract"        
#> [47] "Left Striato-Occipital Tract"              
#> [48] "Right Striato-Occipital Tract"             
#> [49] "Left Striato-Parietal Tract"               
#> [50] "Right Striato-Parietal Tract"              
#> [51] "Left Striato-Postcentral Tract"            
#> [52] "Right Striato-Postcentral Tract"           
#> [53] "Left Striato-Precentral Tract"             
#> [54] "Right Striato-Precentral Tract"            
#> [55] "Left Striato-Prefrontal Tract"             
#> [56] "Right Striato-Prefrontal Tract"            
#> [57] "Left Striato-Premotor Tract"               
#> [58] "Right Striato-Premotor Tract"              
#> [59] "Left Thalamo-Occipital Tract"              
#> [60] "Right Thalamo-Occipital Tract"             
#> [61] "Left Thalamo-Parietal Tract"               
#> [62] "Right Thalamo-Parietal Tract"              
#> [63] "Left Thalamo-Postcentral Tract"            
#> [64] "Right Thalamo-Postcentral Tract"           
#> [65] "Left Thalamo-Precentral Tract"             
#> [66] "Right Thalamo-Precentral Tract"            
#> [67] "Left Thalamo-Prefrontal Tract"             
#> [68] "Right Thalamo-Prefrontal Tract"            
#> [69] "Left Thalamo-Premotor Tract"               
#> [70] "Right Thalamo-Premotor Tract"              
#> [71] "Left Uncinate Fasciculus"                  
#> [72] "Right Uncinate Fasciculus"