mascot: Macroscale Structural Connectomes via Tractography

A data package providing access to macroscale structural connectomes of the human brain derived from diffusion MRI via tractography. Includes two datasets: (1) the HCP1065 population-averaged tractography atlas (Yeh 2022, CC BY-SA 4.0) and (2) per-subject TractSeg white matter bundle segmentations from 105 HCP Young Adult subjects (Wasserthal et al. 2018, CC BY-NC 4.0 – non-commercial use only). Both datasets originate from the Human Connectome Project and are subject to the applicable HCP data-use terms. Source TRK files have been converted to R fiber::bundle objects (.rds); no spatial modifications were made to the streamline data.

See also

Author

Maintainer: Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Authors:

Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Other contributors:

Fang-Cheng Yeh (Author of the HCP1065 atlas) [contributor]

Sandip Panesar (Author of the HCP1065 atlas) [contributor]

David Fernandes (Author of the HCP1065 atlas) [contributor]

Antonio Meola (Author of the HCP1065 atlas) [contributor]

Masanori Yoshino (Author of the HCP1065 atlas) [contributor]

Juan C. Fernandez-Miranda (Author of the HCP1065 atlas) [contributor]

Jean M. Vettel (Author of the HCP1065 atlas) [contributor]

Timothy Verstynen (Author of the HCP1065 atlas) [contributor]

Jakob Wasserthal (Author of the TractSeg dataset (Zenodo DOI: 10.5281/zenodo.1477956)) [contributor]

Peter Neher (Author of the TractSeg dataset (Zenodo DOI: 10.5281/zenodo.1477956)) [contributor]

Klaus H. Maier-Hein (Author of the TractSeg dataset (Zenodo DOI: 10.5281/zenodo.1477956)) [contributor]